Fix sendfile operation on macos
Make sure `op_ctx` to do a reset at scope exist receives own independent object to call reset on it.
I'm not sure of the exact cause but maybe llvm on macos takes a different approach
handling lifetime of the object captured when lambda finishes it's scope.
Copying the context object to reset lambda capture does the job regardles of lifetime of
initial context object.